pyproject.toml
[project]
name = "smart-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.1",
    "gradio>=6.9.0",
    "groq>=1.1.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "uvicorn>=0.41.0",
]
